requirements.txt
python-telegram-bot[socks]==20.7
openai>=1.11
python-dotenv>=1.0
Flask>=2.3